remove tcconfig from servers

+++ src/main/java/midnight/filefix/FileFix.java
@@ -19,12 +19,12 @@ import cpw.mods.fml.common.event.FMLPostInitializationEvent;
 
 @Mod(modid = FileFix.MODID, name = FileFix.MODNAME, version = FileFix.VERSION, dependencies = FileFix.DEPS)
 public class FileFix {
-	
+
 	public static final String MODID = "blightfallfilefix";
 	public static final String MODNAME = "BlightfallFileFixer";
 	public static final String VERSION = "1.0.0";
 	public static final String DEPS = "";
-	
+
 	@Mod.EventHandler
 	// public void preInit(FMLPreInitializationEvent event) {
 	public void postInit(FMLPostInitializationEvent event) {
@@ -40,7 +40,7 @@ public class FileFix {
 			catch (final IOException ex) {
 				ex.printStackTrace();
 			}
-
+			
 			try {
 				check.createNewFile();
 				final BufferedWriter bw = new BufferedWriter(new FileWriter(check));
@@ -56,17 +56,17 @@ public class FileFix {
 			System.out.println("Old version matches current version, skipping.");
 		}
 	}
-
+	
 	private void fixFiles() throws IOException {
 		this.purgeTerrainControlConfigs();
 		this.fixDim112();
 	}
-	
+
 	private void fixDim112() throws IOException {
 		InputStream zipfile;
-		
+
 		zipfile = this.getClass().getResourceAsStream("/midnight/filefix/embeddedfiles/r.-2.-3.mca.zip");
-		
+
 		final ZipInputStream zipIn = new ZipInputStream(zipfile);
 		final ZipEntry entry = zipIn.getNextEntry();
 		final String[] pathnames = { "flans", "DIM-112", "region", entry.getName() };
@@ -75,7 +75,7 @@ public class FileFix {
 		zipIn.closeEntry();
 		zipIn.close();
 	}
-	
+
 	private void extractFile(ZipInputStream zipIn, String filePath) throws IOException {
 		System.out.println("Extracting file: " + filePath);
 		final BufferedOutputStream bos = new BufferedOutputStream(new FileOutputStream(filePath));
@@ -87,28 +87,36 @@ public class FileFix {
 		bos.close();
 		System.out.println("Extracted file: " + filePath);
 	}
-
+	
 	private void purgeTerrainControlConfigs() {
 		final File saves = new File("saves");
-		for (final File directory : saves.listFiles()) {
-			if (directory.isDirectory()) {
-				final File tc = new File(new File(directory, "TerrainControl"), "WorldBiomes");
-				try {
-					System.out.println("Deleting " + tc.getCanonicalPath());
-				}
-				catch (final IOException e1) {
-					System.out.println("Failed to delete " + tc.getName());
-				}
-				try {
-					FileUtils.deleteDirectory(tc);
-				}
-				catch (final IOException e) {
-					System.out.println("Failed to delete " + tc.getName());
+		if (saves.exists() && saves.isDirectory()) {
+			for (final File directory : saves.listFiles()) {
+				if (directory.isDirectory()) {
+					final File tc = new File(new File(directory, "TerrainControl"), "WorldBiomes");
+					try {
+						System.out.println("Deleting " + tc.getCanonicalPath());
+						FileUtils.deleteDirectory(tc);
+					}
+					catch (final IOException e) {
+						System.out.println("Failed to delete " + tc.getName());
+					}
 				}
 			}
 		}
+		final File world = new File("world");
+		if (world.exists() && world.isDirectory()) {
+			final File tc = new File(new File(world, "TerrainControl"), "WorldBiomes");
+			try {
+				System.out.println("Deleting " + tc.getCanonicalPath());
+				FileUtils.deleteDirectory(tc);
+			}
+			catch (final IOException e) {
+				System.out.println("Failed to delete " + tc.getName());
+			}
+		}
 	}
-	
+
 	private String getFileFixerVersion(File file) {
 		try {
 			if (file.exists()) {
@@ -121,5 +129,5 @@ public class FileFix {
 			return "0.0.0";
 		}
 	}
-
+	
 }
